Refactor tests to follow existing convention
